pyproject.toml
[project]
name = "django-weather-reminder"
version = "0.1.0"
description = "A weather notification service with city subscriptions"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"

authors = [
    {name = "DjangoWeatherReminder Team"}
]

keywords = ["django", "weather", "notifications", "celery", "api"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Framework :: Django",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

# Dependency groups
[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.3.0",
    "ipython>=8.20.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

test = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
]

# Main dependencies
dependencies = [
    # Django core
    "django>=5.0.0,<6.0.0",

    # Django REST Framework
    "djangorestframework>=3.15.0",

    # JWT Authentication
    "djangorestframework-simplejwt>=5.3.0",

    # OAuth Authentication
    "django-allauth>=0.57.0",

    # Celery for background tasks
    "celery>=5.3.0",
    "celery[redis]>=5.3.0",

    # Redis client
    "redis>=5.0.0",
    "django-redis>=5.4.0",

    # PostgreSQL adapter
    "psycopg2-binary>=2.9.9",

    # Database URL parser
    "dj-database-url>=2.1.0",

    # Environment variables
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
# Package discovery - find all packages in src/
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
# Dependency groups are defined in [project.optional-dependencies]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
testpaths = ["tests"]

